+++ src/app/_models/query.ts
@@ -4,7 +4,7 @@ export class Query {
     id: number;
     name: string;
     query: string;
-    user_id: number;
+    dashboard_id: number;
     data_source_id: number;
     data_source: DataSource;
 }