+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Issue Fixes
 * [#105](https://github.com/upfrontsystems/izinto-angular/issues/105) fix wind arrow direction
 * [#164](https://github.com/upfrontsystems/izinto-angular/issues/164) allow queries from iframe
 * [#173](https://github.com/upfrontsystems/izinto-angular/issues/173) add Query model and page
+* [#179](https://github.com/upfrontsystems/izinto-angular/issues/179) Link queries per dashboard
 
 
 ## 0.6